repos: rework error management
First step in reworking error handling, this creates a top-level Error
type and matching Result type, the intent is to simplify code by
allowing easier short-circuiting, and to map to various status codes in
the handler layer (to come later).

While we're add it, add the infra necessary to properly log.