handlers: rework error management
Step 2 in reworking errors, this changes fallible handlers to use the
top-level Error type, and reworks the implementations to use '?' for
short-circuiting.

It also implements `actix_web::ResponseError` to generate JSON error
responses for the frontend. The implementation of `status_code()` can be
adjusted to generate appropriate status codes based on error (to come).