user: allocate passkey on user registration
A `passkey` will be encoded in the users' snatched torrent file, and is
used to uniquely identify them to the tracker.  Make use of a 128-bit
passkey, should give plenty of entropy to avoid bruteforce guessing
attacks.

Postgres does not have a native u128 type, so instead marshal the
generated 128-bit type into two 64-bit components, each inserted as an
upper and lower pair.

While we're here, default the `settings` column to an empty dict to
facilitate easier user insertions.