env: rename .env to .env.example and discontinue use of .env.local
Make consistent use of a single .env file as housing runtime parameters
for Arcadia.

While we're here, drop the unmaintained `dotenv` and prefer the
maintained `dotenvy`, which was already in the dependency tree and is
actively maintained.

Note that existing users will likely need to rename their `.env.local`
file to `.env` in order to continue working.