backend: make clippy green
Clippy is a rust tool that does some basic static analysis.  Run it, and
ensure that the backend comes out error and warning free.

In the process, this identified that the `category_enum` had both
`Single` and `SingleCategory` members, but on the backend, they were
being aliased.  Just opt to drop the latter.

Also, ensure clippy will run in CI.