torrent: store torrents in db and insert inject user passkey
Strip the uploaded torrent of all but it's `info_dict`, and store it in
the database instead of storing the full torrent into a separate place.
This should hopefully allow maintaining consistency easier.

When a user downloads a torrent, inject their per-user passkey, and
inject a URL into the comment field. This will likely need some
re-visiting when we figure out exactly what URL should be injected for
front-end viewing.

As part of this, some bugs were discovered in bip_metainfo, vendor it
with bug fixes, until it gets replaced.