backend: use actix-web-httpauth middleware
Clean-up the auth mechanisms to use the actix-provided auth middleware.

This enables all API endpoints to be authenticated using bearer token
(with exception to `/login` and `/register` end points), even if they
don't explicitly make use of the previous `User` extractor.

In addition, this also introduces a `UserId` type and extractor which
will provide only the current user id, which is sufficient for most all
the handlers (which have been updated to use it).  This should prevent
pulling the full `User` record on each endpoint hit.