tracker: fix passkey in downloaded torrent
Postgres only supports signed integer types, which is why we always
get a signed type back.  And yet, we interpret the bits as unsigned (or
at least should have been!).

So, sign-extension strikes again.  If the lower portion of the passkey
had the highest-order bit set, when casted to u128, all of the upper
bits are set to 1, meaning the upper bits are wiped.

Fix this just by ensuring the bits are unsigned before extension.