tracker: fix peer list in announce
The announce response from the tracker needs to ensure that the "peers"
field in the dict encodes the peers into a (byte)string.  Previously,
the tracker was instead encoding the peers as a list of 6-byte strings,
but that was incorrect, which caused torrent clients to ignore the peer
list.

With this fix, we can actually connect peers together :-)